Moved Canvas_TicTacToe
